--- packages/screwdriver-js/src/data-structure/List.ts
@@ -1,7 +1,0 @@
-import { Collection } from './Collection';
-// 列表的接口
-export interface List<A> extends Collection<A> {
-    get(idx: number): A;
-}
-
-export default List;
