--- packages/screwdriver-js/src/global/index.ts
@@ -1,1 +1,0 @@
-import './Array';
