--- packages/screwdriver-js/src/map.ts
@@ -1,9 +1,0 @@
-import { Mapper } from "./fantasy/Mapper";
-import { Functor } from "./fantasy/Functor";
-
-export function map<A, B>(mapper: Mapper<A, B>, fa: Functor<A>): Functor<B> {
-    return fa.map(mapper);
-}
-
-
-export default map;
