+++ packages/screwdriver-js/src/type/ArrayList.ts
@@ -1,7 +1,32 @@
-import { AbstractFunctor } from '../fantasy/Functor';
-import Mapper from '../fantasy/Mapper';
-import { List } from './List';
-import { AbstractList } from './AbstractList';
+import { Mapper, Collection } from "..";
+
+// 列表的接口
+export interface List<A> extends Collection<A> {
+  get(idx: number): A;
+}
+
+abstract class AbstractList<A> implements List<A> {
+  abstract add(item: A): List<A>;
+  abstract size(): number;
+  abstract get(idx: number): A;
+
+
+  abstract map<B>(mapper: Mapper<A, B>): List<B>;
+
+  'fantasy-land/map'<B>(mapper: Mapper<A, B>): List<B> {
+      return this.map(mapper);
+  }
+
+
+  // 抽象类中的粗劣实现，具体类重新实现覆盖一次
+  toArray(): Array<A> {
+      const arr: Array<A> = [];
+      for (let idx = 0; idx < this.size(); idx++) {
+          arr.push(this.get(idx));
+      }
+      return arr;
+  }
+}
 
 export class ArrayList<A> extends AbstractList<A> {
     // 使用数组来存储List的数据
