+++ packages/screwdriver-test/README.md
@@ -1,11 +1,9 @@
 # `screwdriver-test`
 
-> TODO: description
 
 ## Usage
 
 ```
-const screwdriverTest = require('screwdriver-test');
 
-// TODO: DEMONSTRATE API
+
 ```
