+++ packages/screwdriver-test/__tests__/List.test.ts
@@ -1,7 +1,6 @@
-import * as sd from 'screwdriver-js';
-import ArrayList from 'screwdriver-js/src/data-structure/ArrayList';
+import { ArrayList, map } from "screwdriver-js";
 
 test("test map for List22", () => {
     let list = ArrayList.fromArray([1, 2]);
-    expect(sd.map(a => a as number + 1, list)).toEqual(ArrayList.fromArray([2, 3]));
+    expect(map(a => a as number + 1, list)).toEqual(ArrayList.fromArray([2, 3]));
 });
