+++ packages/screwdriver-test/__tests__/add.test.ts
@@ -1,4 +1,4 @@
-import { add } from 'screwdriver-js';
+import { add } from "screwdriver-js";
 
 test("test add 2 and 3", () => {
    expect(add(2,3)).toBe(5);
