--- packages/screwdriver-test/__tests__/index.test.ts
@@ -1,5 +1,0 @@
-import * as sd from 'screwdriver-js'
-
-test("api counts", () => {
-    // expect(Object.keys(sd).length).toBe(2);
-})
